Enable port 8778
This port was used by Jolokia, and enabled by the jolokia module.
Add it to the top-level image YAMLs in order to keep it open for
backwards compatibility — a user may wish to re-add Jolokia to
the images via S2I but opening the port cannot be easily achieved
by that method.

Signed-off-by: Jonathan Dowland <jdowland@redhat.com>