WIP: Turn off negative DNS caching (OPENJDK-100) for ubi8/openjdk-8-runtime
Extend the code that disables DNS name caching to the
ubi8/openjdk-8-runtime image.

Note that we cannot extend this for the ubi8/openjdk-11-runtime image
yet, because that image (and java-11-openjdk{,-headless} packages) lack
a java.security file.

I actually think we should move this code to the JDK/JRE cekit modules.

Signed-off-by: Jonathan Dowland <jdowland@redhat.com>