Move DNS caching tuning to JDK/JRE modules
Rather than handle this in the run module (which should really be
restricted to matters relevant to the run-java script), move the code
that tunes the java.security file to the OpenJDK JDK and "JRE" modules
that handle installing (and configuring) the OpenJDK packages.

Extend it to the runtime images.

Signed-off-by: Jonathan Dowland <jdowland@redhat.com>