Move nss_wrapper stuff out of run module
Introduce a new OPENJDK-312-nss-wrapper module that encapsulates all
the installation and configuration of nss_wrapper needed for backwards
compatibility of the builder images on OCP 3.11 after the fix for
OPENJDK-312 was merged.

Leverage the new hooks.d scheme in the run module so that it is not
tightly coupled to this one.

Signed-off-by: Jonathan Dowland <jdowland@redhat.com>